update README newDebian.sh Changes to be committed: 	modified:   README-zh_CN.md 	modified:   README.md 	modified:   newDebian.sh

+++ newDebian.sh
@@ -2,9 +2,17 @@
 
 # change source
 # https://github.com/oooldking/script
+
+read -p "Whether to change the source? [Y/n]" varify
+
+if [ $varify = 'y' ]||[ $varify = 'Y' ]; then
+echo "Please select fastly(default), cn, 163 or aliyun."
+read source
 wget    git.io/superupdate.sh
-bash    superupdate.sh
+bash    superupdate.sh $source
 rm      superupdate.sh
+fi
+
 
 # update apt
 apt update
